fix (changed karma test spec files after rename material module)
