+++ src/app/view/dashboard/project/project-resources/project-resources.component.spec.ts
@@ -1,6 +1,6 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {AppModule} from '../../../../app.module';
-import {AppMaterialModule} from '../../../../app-material.module';
+import {AppMaterialModule} from '../../../../app-material-module';
 import {AppRoutingModule} from '../../../../app-routing.module';
 import {ProjectResourcesComponent} from './project-resources.component';
 import {ResourceTypesService} from '../../../../model/services/resource-types.service';
