+++ src/app/view/modules/form/advanced-resource-class/advanced-resource-class.component.spec.ts
@@ -1,6 +1,6 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {AppModule} from '../../../../app.module';
-import {AppMaterialModule} from '../../../../app-material.module';
+import {AppMaterialModule} from '../../../../app-material-module';
 import {AppRoutingModule} from '../../../../app-routing.module';
 import {AdvancedResourceClassComponent} from './advanced-resource-class.component';
 import {BaseOntologyService} from 'app/model/services/base-ontology.service';
