+++ src/app/view/modules/form/resource-class-form/resource-class-form.component.spec.ts
@@ -1,6 +1,6 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {AppModule} from '../../../../app.module';
-import {AppMaterialModule} from '../../../../app-material.module';
+import {AppMaterialModule} from '../../../../app-material-module';
 import {AppRoutingModule} from '../../../../app-routing.module';
 import {ResourceClassFormComponent} from './resource-class-form.component';
 import {BaseOntologyService} from '../../../../model/services/base-ontology.service';
