+++ src/app/view/modules/header/header-toolbar/header-toolbar.component.spec.ts
@@ -1,6 +1,6 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {AppModule} from '../../../../app.module';
-import {AppMaterialModule} from '../../../../app-material.module';
+import {AppMaterialModule} from '../../../../app-material-module';
 import {AppRoutingModule} from '../../../../app-routing.module';
 import {HeaderToolbarComponent} from './header-toolbar.component';
 import {SessionService} from '../../../../model/services/session.service';
