+++ src/app/view/modules/listing/projects-list/projects-list.component.spec.ts
@@ -1,6 +1,6 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {AppModule} from '../../../../app.module';
-import {AppMaterialModule} from '../../../../app-material.module';
+import {AppMaterialModule} from '../../../../app-material-module';
 import {AppRoutingModule} from '../../../../app-routing.module';
 import {ProjectsListComponent} from './projects-list.component';
 import {ProjectsService} from '../../../../model/services/projects.service';
